Refactor database actions and user management; remove unused components and pages
