--- src/components/StuffItem.tsx
@@ -1,16 +1,0 @@
-import { Stuff } from '@prisma/client';
-import Link from 'next/link';
-
-/* Renders a single row in the List Stuff table. See list/page.tsx. */
-const StuffItem = ({ name, quantity, condition, id }: Stuff) => (
-  <tr>
-    <td>{name}</td>
-    <td>{quantity}</td>
-    <td>{condition}</td>
-    <td>
-      <Link href={`/edit/${id}`}>Edit</Link>
-    </td>
-  </tr>
-);
-
-export default StuffItem;
