--- src/components/StuffItemAdmin.tsx
@@ -1,16 +1,0 @@
-import { Stuff } from '@prisma/client';
-
-/* Renders a single row in the List Stuff table. See list/page.tsx. */
-const StuffItemAdmin = ({ name, quantity, condition, owner, id }: Stuff) => (
-  <tr>
-    <td>{name}</td>
-    <td>{quantity}</td>
-    <td>{condition}</td>
-    <td>{owner}</td>
-    <td>
-      <a href={`/edit/${id}`}>Edit</a>
-    </td>
-  </tr>
-);
-
-export default StuffItemAdmin;
