+++ src/lib/dbUserActions.ts
@@ -1,0 +1,34 @@
+'use server';
+
+import { hash } from 'bcrypt';
+import { prisma } from './prisma';
+
+/**
+ * Creates a new user in the database.
+ * @param credentials, an object with the following properties: email, password.
+ */
+export async function createUser(credentials: { email: string; password: string }) {
+  // console.log(`createUser data: ${JSON.stringify(credentials, null, 2)}`);
+  const password = await hash(credentials.password, 10);
+  await prisma.user.create({
+    data: {
+      email: credentials.email,
+      password,
+    },
+  });
+}
+
+/**
+ * Changes the password of an existing user in the database.
+ * @param credentials, an object with the following properties: email, password.
+ */
+export async function changePassword(credentials: { email: string; password: string }) {
+  // console.log(`changePassword data: ${JSON.stringify(credentials, null, 2)}`);
+  const password = await hash(credentials.password, 10);
+  await prisma.user.update({
+    where: { email: credentials.email },
+    data: {
+      password,
+    },
+  });
+}
